Adding udp crc check back
I've been testing and analyzing and am not seeing any issues with the
UDP crc checks as it relates to the camera manager.  All messages from
qground to control the gimbal and camera functions are working.  And the
reporting of gimbal status, available sd card space, etc. are all
working.  As such I've added the udp check back in with one error log
(which I'm never seeing in the output right now) to catch it if it does
start to become a problem.